Machine Learning w. Text Analytics
This is a project using the Real & Fake News Dataset which can be found on the Kaggle website. The research question on this page asks if there is an algorithm that can correctly predict Fake news from Real news. I used Machine learning Algorithms along with Sentiment Analysis to answer this question.